refactor - index접근방식 수정
개발자가 볼테니 항상 0시작인덱스로접근.
