+++ src/main/java/nextstep/Ladder.java
@@ -3,6 +3,6 @@ package nextstep;
 import java.util.List;
 
 public abstract class Ladder {
-    abstract String ladderString(int targetVertical);
-    abstract void lineMark(int targetVertical, int targetHorizon);
+    public abstract String ladderString(int targetVertical);
+    public abstract void lineMark(int targetVertical, int targetHorizon);
 }
