+++ src/main/java/nextstep/Lines.java
@@ -7,10 +7,15 @@ public class Lines {
     private final List<Line> lines;
 
     public Lines(int vertical, int countOfPerson) {
+        System.out.println("countOfPerson = " + countOfPerson);
         this.lines = new ArrayList<>();
-        for (int i = 0; i <= vertical; i++) {
+        for (int i = 0; i < vertical; i++) {
             this.lines.add(new Line(countOfPerson));
         }
+    }
+
+    public void mark(int targetVertical) {
+        this.lines.get(targetVertical).mark();
     }
 
     public void mark(int targetVertical, int targetHorizon) {
