+++ src/main/java/nextstep/view/ResultView.java
@@ -2,20 +2,11 @@ package nextstep.view;
 
 import nextstep.OnlineLadder;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
 public class ResultView implements ConsoleView {
-    private List<Integer> beforeHeight;
     private final int ladderCount;
 
     public ResultView(int ladderCount) {
         this.ladderCount = ladderCount;
-        this.beforeHeight = new ArrayList<>();
-        for (int i = 0; i < ladderCount; i++) {
-            this.beforeHeight.add(0);
-        }
     }
 
     @Override
@@ -34,7 +25,6 @@ public class ResultView implements ConsoleView {
         }
         System.out.println();
         for (int i = 0; i < this.ladderCount; i++) {
-//            ladder.move();
             this.print(ladder, i);
         }
     }
