+++ src/test/java/nextstep/ladder/CoordinateTest.java
@@ -1,5 +1,6 @@
-package nextstep;
+package nextstep.ladder;
 
+import nextstep.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
